Moved some more and wrote some comments
