+++ crates/browser/src/tab/mod.rs
@@ -1,3 +1,16 @@
+//! Tab-level orchestration and streaming state.
+//!
+//! Invariants:
+//! - `nav_gen` is the navigation/request generation counter. All streaming
+//!   events are gated through `is_current` so stale events from previous
+//!   generations are ignored.
+//! - `PageState::pending_count()` is the single source of truth for whether
+//!   the tab is still loading; `loading` is derived from it and only tracks
+//!   user-visible state.
+//! - Each `Tab` owns its `PageState`, `ResourceManager`, and `InteractionState`.
+//!   There is no cross-tab sharing of DOM, resources, or input state; any
+//!   shared work must go through the bus/runtime layers.
+
 use crate::page::PageState;
 use crate::resources::ResourceManager;
 use app_api::RepaintHandle;
@@ -8,9 +21,9 @@ use std::sync::mpsc;
 use html::{Node, dom_utils::assign_node_ids};
 
 use core_types::{RequestId, ResourceKind, TabId};
-use css::parse_color;
 
 mod discovery;
+mod dom_style;
 mod nav;
 mod ui;
 
@@ -315,56 +328,6 @@ impl Tab {
             repaint.request_now();
         }
     }
-
-    pub fn inherited_color(node: &Node, ancestors: &[Node]) -> (u8, u8, u8, u8) {
-        fn find_on(node: &Node) -> Option<(u8, u8, u8, u8)> {
-            match node {
-                Node::Element { style, .. } => style
-                    .iter()
-                    .find(|(k, _)| k.eq_ignore_ascii_case("color"))
-                    .and_then(|(_, v)| parse_color(v)),
-                _ => None,
-            }
-        }
-        if let Some(c) = find_on(node) {
-            return c;
-        }
-        for a in ancestors {
-            if let Some(c) = find_on(a) {
-                return c;
-            }
-        }
-        (0, 0, 0, 255) // default black
-    }
-
-    pub fn page_background(dom: &Node) -> Option<(u8, u8, u8, u8)> {
-        fn from_element(node: &Node, want: &str) -> Option<(u8, u8, u8, u8)> {
-            match node {
-                Node::Element { name, style, .. } if name.eq_ignore_ascii_case(want) => style
-                    .iter()
-                    .find(|(k, _)| k.eq_ignore_ascii_case("background-color"))
-                    .and_then(|(_, v)| parse_color(v)),
-                _ => None,
-            }
-        }
-        if let Node::Document { children, .. } = dom {
-            for c in children {
-                if let Some(c1) = from_element(c, "html") {
-                    return Some(c1);
-                }
-                if let Node::Element {
-                    children: html_kids,
-                    ..
-                } = c
-                {
-                    for k in html_kids {
-                        if let Some(c2) = from_element(k, "body") {
-                            return Some(c2);
-                        }
-                    }
-                }
-            }
-        }
-        None
-    }
 }
+
+pub use dom_style::{inherited_color, page_background};
